file restructure, so easier to understand
