+++ Frontend/src/components/mentee/mentee-profile/MenteeProfile.jsx
@@ -1,9 +1,10 @@
 import {useContext, useEffect, useState} from 'react'
-import { UserContext } from '../../UserContext.jsx';
+import { UserContext } from '../../../UserContext.jsx';
 import { useNavigate, Link } from 'react-router-dom';
-import './MentorProfile.css'
-import MentorProfileModal from './MentorProfileModal.jsx';
-import config from '../../../config.js';
+
+import './MenteeProfile.css'
+import MenteeProfileModal from './MenteeProfileModal.jsx';
+import config from '../../../../config.js';
 
 const PLACEHOLDER = "https://ralfvanveen.com/wp-content/uploads/2021/06/Placeholder-_-Glossary.svg";
 
@@ -53,40 +54,27 @@ const skillsEnum = Object.freeze({
 
 const skillsList = Object.values(skillsEnum)
 
-const experienceMappingReverse = {
-    1: '0-2',
-    2: '2-5',
-    3: '5-10',
-    4: '10+',
-    5: '20+',
-  };
-
-function MentorProfile() {
+function MenteeProfile() {
     const { user } = useContext(UserContext)
     const [selectedSkills, setSelectedSkills] = useState([]);
     const [dropdownOpen, setDropdownOpen] = useState(false);
     const [isModalOpen, setIsModalOpen] = useState(false)
     const navigate = useNavigate();
     // const logoApiKey = process.env.REACT_APP_LOGO_API_KEY; getting error process not defined
-    
-    const fetchLogo = () => {
-        // use api to fetch logo
-    };
 
     const [userData, setUserData] = useState({
         name: "Loading",
         profileImageUrl: PLACEHOLDER,
-        industry: '',
-        company: '',
-        work_role: '',
-        years_experience: '',
+        bio: '',
+        major: '',
+        career_goals: '',
         school: '',
         skills: ''
     });
 
-    const fetchMentorData = () => {
+    const fetchMenteeData = () => {
         if (user && user.id) {
-          fetch(`${config.apiBaseUrl}/mentors/${user.id}`)
+          fetch(`${config.apiBaseUrl}/mentees/${user.id}`)
             .then(response => {
               if (!response.ok) {
                 throw new Error(`HTTP status ${response.status}`);
@@ -97,17 +85,15 @@ function MentorProfile() {
               setUserData({
                 name: user.name,
                 profileImageUrl: user.profileImageUrl || PLACEHOLDER,
-                industry: data.mentor.industry,
-                company: data.mentor.company,
-                work_role: data.mentor.work_role,
-                years_experience: experienceMappingReverse[data.mentor.years_experience] || '',
-                school: data.mentor.school,
-                bio: data.mentor.bio,
-                skills: data.mentor.skills
+                major: data.mentee.major,
+                school: data.mentee.school,
+                bio: data.mentee.bio,
+                career_goals: data.mentee.career_goals,
+                skills: data.mentee.skills
               });
             })
             .catch(error => {
-              console.error('Error fetching mentor data:', error);
+              console.error('Error fetching mentee data:', error);
               setUserData({
                 name: 'Failed to load user data',
                 profileImageUrl: PLACEHOLDER
@@ -116,9 +102,8 @@ function MentorProfile() {
         }
       };
       
-
     useEffect(() => {
-        fetchMentorData();
+        fetchMenteeData();
     }, [user]);
 
     const handleCheckboxChange = (skill) => {
@@ -137,19 +122,19 @@ function MentorProfile() {
         setIsModalOpen(!isModalOpen);
     }
 
-
     return(
         <>
         <div className="mp-header">
             <h1>CareerCompass</h1>
             <div className='mp-nav'>
-                <Link to="/mentor-dashboard">Dashboard</Link>
+                <Link to="/mentee-dashboard">Dashboard</Link>
+                <Link to="/matching">Find Mentors</Link>
             </div>
         </div>
         <div className='mp-container'>
             <div className='mp-top'>
                 <div className='mp-top-left'>
-                    <h1>Mentor Profile</h1>
+                    <h1>Mentee Profile</h1>
                 </div>
                 <div className='mp-top-right'>
                     <button onClick={handleModalToggle}>Edit</button>
@@ -161,21 +146,16 @@ function MentorProfile() {
                     <h3>{userData.name}</h3>
                 </div>
                 <div className='mp-right'>
-                    <p>Industry: {userData.industry}</p>
-                    <div className='mp-right-company'>
-                        <p>Company: {userData.company}</p>
-                        <img src={`https://img.logo.dev/${userData.company}.com?token=pk_DCOxK2D7TA68fkEDQQ2_fQ`} />
-                    </div>
-                    <p>Role: {userData.work_role}</p>
-                    <p>Years of Experience: {userData.years_experience} years</p>
                     <p>School: {userData.school}</p>
+                    <p>Major: {userData.major}</p>
+                    <p>Career Goals: {userData.career_goals}</p>
                     <p>Skills: {userData.skills}</p>
                     <p>Bio: {userData.bio}</p>
                 </div>
             </div>
             {isModalOpen && (
-                    <MentorProfileModal
-                        mentorData={userData}
+                    <MenteeProfileModal
+                        menteeData={userData}
                         handleDropdownToggle={handleDropdownToggle}
                         dropdownOpen={dropdownOpen}
                         selectedSkills={selectedSkills}
@@ -183,14 +163,14 @@ function MentorProfile() {
                         skillsList={skillsList}
                         closeModal={() => {
                             handleModalToggle();
-                            fetchMentorData();
+                            fetchMenteeData();
                         }}
                     />
                 )}
-            <button onClick={() => navigate('/mentor-dashboard')} id='save-button'>Save</button>
+            <button onClick={() => navigate('/mentee-dashboard')} id='save-button'>Save</button>
         </div>
         </>
     )
 }
 
-export default MentorProfile
+export default MenteeProfile
